+++ src/index.ts
@@ -1,27 +1,6 @@
-import React from "React";
+import Count from "Components/Count";
+import React from "../React";
 
-const element = React.createElement("div", {
-  className: "container",
-  children: [
-    React.createElement("span", {
-      className: "count",
-      children: "0",
-    }),
-    React.createElement("div", {
-      className: "btn-group",
-      children: [
-        React.createElement("button", {
-          children: [React.createElement("strong", { children: "-" })],
-        }),
-        React.createElement("button", {
-          children: [React.createElement("strong", { children: "RESET" })],
-        }),
-        React.createElement("button", {
-          children: [React.createElement("strong", { children: "+" })],
-        }),
-      ],
-    }),
-  ],
-});
+const element = Count();
 
 React.render(element, document.querySelector("#root"));
